Units Attack System implemented Fog of War feature implemented Debug feature implmented

+++ Assets/Warcraft/Components/Player/PlayerUnitsComponent.cs
@@ -24,7 +24,7 @@ namespace Warcraft.Components {
 
         public void AddUnit(int unitTypeId) {
 
-            this.units |= (ulong)(1 << unitTypeId);
+            this.units |= (ulong)(long)(1 << unitTypeId);
 
         }
